# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "McDee Platform API",
        "description": "Multifunctional Marketplace & Services API",
        "contact": {
            "email": "support@mcdee.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get current user's cart",
                "operationId": "0652f94dde92e36bbb3c461356ce1d7b",
                "responses": {
                    "200": {
                        "description": "Cart retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add item to cart",
                "operationId": "408de867ef91c8a1c1298be6461a5e3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_type",
                                    "item_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "item_type": {
                                        "type": "string",
                                        "example": "product"
                                    },
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added to cart"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/remove/{item_id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove item from cart",
                "operationId": "384f6db375db64ec88cc644b8a8d023e",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed from cart"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear entire cart",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "Cart cleared"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/orders": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get current user's orders",
                "operationId": "08317b752b9a6625cb3123949ef4022b",
                "responses": {
                    "200": {
                        "description": "List of user's orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Place a new order",
                "operationId": "d690d6834318e9b064df441e854de8b9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "items"
                                ],
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order placed successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/vendor": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders for vendor's products",
                "operationId": "0ae6b9e4a682706af88fd1e56854dbb9",
                "responses": {
                    "200": {
                        "description": "List of vendor's orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/orders/{id}/status": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update order status (e.g., processing, completed)",
                "operationId": "a6abe7e1fd2257d370bf832e8075f50c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "processing",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "403": {
                        "description": "Unauthorized or not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/products/{id}/review": {
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Submit a review for a product",
                "operationId": "69f8b1da60e1a74e4ad997e71f34b148",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "rating"
                                ],
                                "properties": {
                                    "rating": {
                                        "type": "integer",
                                        "maximum": 5,
                                        "minimum": 1
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review submitted"
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/settings": {
            "get": {
                "tags": [
                    "Ride Hailing Settings"
                ],
                "summary": "Get ride fare settings",
                "operationId": "fb7f24c8a9d003a265dd0b6b6c79246e",
                "responses": {
                    "200": {
                        "description": "Returns current fare settings"
                    }
                }
            },
            "put": {
                "tags": [
                    "Ride Hailing Settings"
                ],
                "summary": "Update fare settings (admin)",
                "operationId": "23b8084b4a6d4a02d860df48f03cf96b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "base_fare": {
                                        "type": "number"
                                    },
                                    "rate_per_km": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Settings updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}